Create 3 persistent VPCs for dev envs (#31)
To practice bringing EKS clusters up and down, it helps to have
persistent networks to host them.

One of these (dev1) already exists. Here we test setting our nat gateway
and interent gateway booleans to false on an existing network. We want
to be able to do this, becuase these are compute resources that cost
money and aren't needed when there are no clusters or EC2s running in
their VPC.